+++ src/plugins/observe.dark.script.js
@@ -1,0 +1,11 @@
+import svgPlugin from './observe.dark.js';
+
+if (globalThis.gSVG) {
+  globalThis.gSVG.install(svgPlugin);
+}
+
+customElements
+  .whenDefined('g-composer')
+  .then(composer => {
+    composer.install(svgPlugin);
+  });