+++ test/plugins/template.engine/cases/case15.js
@@ -3,10 +3,15 @@ export const description = 'g-bind:class with an object of values (true or false
 
 export function script () {
   const svg = gSVG(document.querySelector('#svg'));
-  svg.render({
-    isWarning : true
-  });
-  document.querySelector('#result').innerHTML = sourceFormat(svg.source());
+  const check = document.querySelector('#update');
+  function update() {
+    svg.render({
+      isWarning : check.checked
+    });
+    document.querySelector('#result').innerHTML = sourceFormat(svg.source());
+  }
+  update();
+  check.addEventListener('change', update);
 }
 
 export default `<svg viewBox="0 0 100 100" id="svg" style="width: 100px; height: 100px">
@@ -22,4 +27,7 @@ export default `<svg viewBox="0 0 100 100" id="svg" style="width: 100px; height:
   <circle cx="25" cy="50" r="20" :class="{warning: isWarning, bordered: true}"></circle>
   <circle cx="75" cy="50" r="20" class="warning" :class="{warning: !isWarning, bordered: true}"></circle>
 </svg>
+<p>
+  <label><input type="checkbox" id="update" checked>isWarning</label>
+</p>
 <pre id="result"></pre>`;