+++ test/plugins/template.engine/cases/case38.js
@@ -1,5 +1,5 @@
 export const title       = '38) .source()';
-export const description = 'the plugin rewrite svg.toSource() method for remove <code>defs g-for</code>';
+export const description = 'the plugin rewrite svg.toSource() method for remove <code>g g-for</code>';
 
 
 export function script () {
@@ -18,7 +18,7 @@ export default `<svg viewBox="0 0 100 100" id="svg" style="width: 100px; height:
   <defs>
     <circle id="test" r="25"></circle>
   </defs>
-  <g g-for="(value, key) of data">
+  <g g-for="(value, key) in data">
     <use
        href = "#test"
         :id = '"circle_" + key'